Make `tidyselect`and `datamasking` safe in aso functions.
When using data masking and tidy selection no global variable with that name is avalible - it only exisits in the context of the current data frame. Therefore, the CRAN check will complain that there is "no global variable definition for" the variable.

In a tidy select context we just use the quoted column name. In a data masking context we must explicitly refere to the current context data fram by writing `.data$variable` when using `variable`.