Add convex optimization support to fct_confint
Use convex optimization to find confidence set of functions of model parameters unless problem is not convex or user specifies otherwise. This solver is expected to be much faster than the previous solver, where a grid of points was projected to the boundary, and is also more accurate, as we are guarantied to get the correct boudaries.

Add cli, snow, and doSNOW to suggests.

Add improved checks of input to fct_confint.