Update input checks in freq_function
Update the checks of weightvar. When checking the content of the weightvar column of normaldata, a check for NAs are added to the existing check of negative weights.

The warning if negative weights or NAs are present is updated to print the index of consecutive rows in the format "a:b".